requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Auth & Security
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Database
supabase==2.27.2
postgrest==2.27.2

# Storage
boto3==1.35.95

# LLM & AI
google-genai>=1.0.0

# Image Processing
Pillow==11.1.0
pymupdf==1.25.2

# Data Processing
pydantic==2.11.7
pydantic-settings==2.7.1

# Utils
python-dateutil==2.9.0
aiofiles==24.1.0
beautifulsoup4==4.12.3

# Logging
structlog==24.4.0

